removed unnecessary benchmark
